backend/ml_models/anomaly_detector/requirements.txt
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
psycopg2-binary==2.9.9
numpy==2.1.3
onnx==1.17.0
skl2onnx==1.17.0
